# services/api-gateway/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api-gateway",
    "description": "Single authenticated ingress over the platform's internal services. Services and AI agents call it with a bearer token to read catalog products, offers and price history, and to curate matches — one consistent, scoped, rate-limited contract.",
    "contact": {
      "name": "Supplier Aggregation Platform",
      "email": "Info_DIT@erc.ua"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/v1/products": {
      "get": {
        "tags": [
          "gateway"
        ],
        "summary": "List supplier products",
        "description": "List supplier products from the catalog, optionally filtered by supplier or the canonical product they map to. Cursor-paginated. Requires scope `catalog:read`.",
        "operationId": "listProducts",
        "parameters": [
          {
            "name": "supplier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by supplier code, e.g. `brain`.",
              "title": "Supplier"
            },
            "description": "Filter by supplier code, e.g. `brain`."
          },
          {
            "name": "canonical_product_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to products mapped to this canonical id (ULID).",
              "title": "Canonical Product Id"
            },
            "description": "Filter to products mapped to this canonical id (ULID)."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous response's next_cursor.",
              "title": "Cursor"
            },
            "description": "Opaque cursor from a previous response's next_cursor."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Max items per page (1-200).",
              "default": 50,
              "title": "Limit"
            },
            "description": "Max items per page (1-200)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Per-principal rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "A downstream service was unreachable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/products/{supplier_product_id}": {
      "get": {
        "tags": [
          "gateway"
        ],
        "summary": "Get one supplier product",
        "description": "Fetch a single supplier product by its id, including the canonical product it maps to (if matched). Requires scope `catalog:read`.",
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "supplier_product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Supplier product id (ULID).",
              "title": "Supplier Product Id"
            },
            "description": "Supplier product id (ULID)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Per-principal rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "A downstream service was unreachable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/products/{supplier_product_id}/offers": {
      "get": {
        "tags": [
          "gateway"
        ],
        "summary": "List offers for a product",
        "description": "List every supplier-account offer (price and availability) for a supplier product. Cursor-paginated. Requires scope `offers:read`.",
        "operationId": "listProductOffers",
        "parameters": [
          {
            "name": "supplier_product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Supplier product id (ULID).",
              "title": "Supplier Product Id"
            },
            "description": "Supplier product id (ULID)."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a previous response's next_cursor.",
              "title": "Cursor"
            },
            "description": "Opaque cursor from a previous response's next_cursor."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Max items per page (1-200).",
              "default": 50,
              "title": "Limit"
            },
            "description": "Max items per page (1-200)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Per-principal rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "A downstream service was unreachable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/products/{supplier_product_id}/best-offer": {
      "get": {
        "tags": [
          "gateway"
        ],
        "summary": "Get the cheapest offer for a product",
        "description": "Return the cheapest current offer (by UAH price) across all supplier accounts for a supplier product. Requires scope `offers:read`.",
        "operationId": "getBestOffer",
        "parameters": [
          {
            "name": "supplier_product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Supplier product id (ULID).",
              "title": "Supplier Product Id"
            },
            "description": "Supplier product id (ULID)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Per-principal rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "A downstream service was unreachable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/v1/curation/stats": {
      "get": {
        "tags": [
          "gateway"
        ],
        "summary": "Curation stats for the operator dashboard",
        "description": "Aggregate counts for the operator dashboard: queue depth, total canonical products, and decisions by action. Requires scope `matching:curate`.",
        "operationId": "getCurationStats",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Per-principal rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "A downstream service was unreachable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Stable URI identifying the error type."
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Short, human-readable summary of the error type."
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "HTTP status code."
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human/LLM-readable explanation with a next step."
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance",
            "description": "URI of the specific occurrence."
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id",
            "description": "Trace id to correlate with telemetry."
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "Problem",
        "description": "RFC 9457 problem+json error body (returned on 4xx/5xx)."
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Service API key or OIDC token, issued/validated by the API gateway."
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This service (behind the API gateway)."
    }
  ],
...
}